Se actualizo SDL2 a SDL3   - Miragracion de SDL2 a SDL3 Se actualizo Imgui a v1.91.9-docking Se actualizo Vulkan Headers, Loader y VVL a v1.4.313 Se actualizo RapidYaml a v0.9.0
Correcion de advertencias
Arreglos menores

+++ include/PatataEngine/StructEngineInfo.hpp
@@ -13,7 +13,7 @@ namespace Patata
 struct EngineInfo
 {
   // Window
-  std::uint32_t WindowCreationFlags = 0;
+  std::uint64_t WindowCreationFlags = 0;
 
   // RaccoonRenderer
   // Vulkan
@@ -40,7 +40,7 @@ struct EngineInfo
   vk::PresentModeKHR VkSwapchainPresentMode;
   vk::Format         VkSwapchainImageColorFormat;
   vk::ColorSpaceKHR  VkSwapchainImageColorSpace;
-  std::uint32_t       VkMinImageCount = 0;
+  std::uint32_t      VkMinImageCount = 0;
 
   // Extensions
   const char ** ppVkLayers                = nullptr;
